Move _default_from_env_vars from env_vars_connector to utils
