+++ pytorch_lightning/trainer/connectors/env_vars_connector.py
@@ -12,29 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from functools import wraps
-from typing import Callable
+from pytorch_lightning.utilities import rank_zero_deprecation
 
-from pytorch_lightning.utilities.argparse import get_init_arguments_and_types, parse_env_variables
+rank_zero_deprecation(
+    "Using `pytorch_lightning.trainer.connectors.env_vars_connector._defaults_from_env_vars` is "
+    "deprecated in v1.6, and will be removed in v1.8. It has been replaced with "
+    "`pytorch_lightning.utilities.argsparse._defaults_from_env_vars`"
+)
 
-
-def _defaults_from_env_vars(fn: Callable) -> Callable:
-    """Decorator for :class:`~pytorch_lightning.trainer.trainer.Trainer` methods for which input arguments should
-    be moved automatically to the correct device."""
-
-    @wraps(fn)
-    def insert_env_defaults(self, *args, **kwargs):
-        cls = self.__class__  # get the class
-        if args:  # inace any args passed move them to kwargs
-            # parse only the argument names
-            cls_arg_names = [arg[0] for arg in get_init_arguments_and_types(cls)]
-            # convert args to kwargs
-            kwargs.update(dict(zip(cls_arg_names, args)))
-        env_variables = vars(parse_env_variables(cls))
-        # update the kwargs by env variables
-        kwargs = dict(list(env_variables.items()) + list(kwargs.items()))
-
-        # all args were already moved to kwargs
-        return fn(self, **kwargs)
-
-    return insert_env_defaults
+from pytorch_lightning.utilities.argparse import _defaults_from_env_vars  # noqa: E402, F401
